refactor(): new structure for controllers and views #41
