+++ src/main/kotlin/module-info.java
@@ -6,6 +6,8 @@ module pl.gawor.taycknerdesktopclient {
 
 
     opens pl.gawor.taycknerdesktopclient to javafx.fxml;
-    opens pl.gawor.taycknerdesktopclient.controller to javafx.fxml;
+    opens pl.gawor.taycknerdesktopclient.controller.dayplanner to javafx.fxml;
+    opens pl.gawor.taycknerdesktopclient.controller.daytracker to javafx.fxml;
+    opens pl.gawor.taycknerdesktopclient.controller.habittracker to javafx.fxml;
     exports pl.gawor.taycknerdesktopclient;
 }
