+++ src/main/kotlin/pl/gawor/taycknerdesktopclient/TaycknerApplication.kt
@@ -7,7 +7,7 @@ import javafx.stage.Stage
 
 class TaycknerApplication : Application() {
     override fun start(stage: Stage) {
-        val fxmlLoader = FXMLLoader(TaycknerApplication::class.java.getResource("view/view.fxml"))
+        val fxmlLoader = FXMLLoader(TaycknerApplication::class.java.getResource("view/habit_tracker/view.fxml"))
         val scene = Scene(fxmlLoader.load(), 1920.0, 1013.0)
         stage.title = "Tayckner"
         stage.scene = scene
