+++ src/main/kotlin/pl/gawor/taycknerdesktopclient/controller/daytracker/ItemActivityController.kt
@@ -1,4 +1,4 @@
-package pl.gawor.taycknerdesktopclient.controller
+package pl.gawor.taycknerdesktopclient.controller.daytracker
 
 import javafx.fxml.FXML
 import javafx.scene.control.Label
