+++ src/main/kotlin/pl/gawor/taycknerdesktopclient/controller/habittracker/ItemHabitController.kt
@@ -1,22 +1,21 @@
-package pl.gawor.taycknerdesktopclient.controller
+package pl.gawor.taycknerdesktopclient.controller.habittracker
 
 import javafx.fxml.FXML
 import javafx.scene.control.Label
 import javafx.scene.layout.VBox
 import pl.gawor.taycknerdesktopclient.controller.Observer.ISubscriber
-import pl.gawor.taycknerdesktopclient.model.Category
+import pl.gawor.taycknerdesktopclient.model.Habit
 
-class ItemCategoryController{
-
+class ItemHabitController {
     @FXML private lateinit var label_name: Label
 
     @FXML private lateinit var label_description: Label
 
     @FXML private lateinit var vbox_root: VBox
 
-    private lateinit var model: Category
+    private lateinit var model: Habit
 
-    fun set(model: Category) {
+    fun set(model: Habit) {
         this.model = model
         label_name.text = model.name
         label_description.text = model.description
@@ -26,24 +25,24 @@ class ItemCategoryController{
     @FXML fun vbox_rootOnMouseClicked() {
         notifySubscribers()
     }
-
     //---// S U B S C R I B E R S
-    private val subscribers: ArrayList<ISubscriber<Category>> = ArrayList()
+    private val subscribers: ArrayList<ISubscriber<Habit>> = ArrayList()
 
     //---// S U B S C R I B E
-    fun subscribe(s: ISubscriber<Category>) {
+    fun subscribe(s: ISubscriber<Habit>) {
         subscribers.add(s)
     }
 
     //---// U N S U B S C R I B E
-    fun unsubscribe(s: ISubscriber<Category>) {
+    fun unsubscribe(s: ISubscriber<Habit>) {
         subscribers.remove(s)
     }
 
     //---// N O T I F Y  S U B S C R I B E R S
-    fun notifySubscribers() {
+    private fun notifySubscribers() {
         for (s in subscribers) {
             s.update(model)
         }
     }
+
 }
