+++ src/main/resources/pl/gawor/taycknerdesktopclient/view/day_planner/item_schedule.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<HBox alignment="CENTER_LEFT" onMouseClicked="#hboxOnMouseClicked" prefHeight="60.0" prefWidth="1000" style="-fx-background-color: #eeeeee;" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.ItemScheduleController">
+<HBox alignment="CENTER_LEFT" onMouseClicked="#hboxOnMouseClicked" prefHeight="60.0" prefWidth="1000" style="-fx-background-color: #eeeeee;" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.dayplanner.ItemScheduleController">
     <padding>
         <Insets left="10.0" right="10.0" />
     </padding>
