+++ src/main/resources/pl/gawor/taycknerdesktopclient/view/day_planner/view.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="1013.0" prefWidth="1920.0" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.Controller">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="1013.0" prefWidth="1920.0" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.dayplanner.Controller">
 
     <HBox prefHeight="1013.0" prefWidth="1920.0">
         <VBox alignment="TOP_CENTER" prefHeight="1013.0" prefWidth="100.0">
@@ -21,7 +21,7 @@
                 <Button fx:id="button_prev" mnemonicParsing="false" onAction="#button_prevOnAction" prefHeight="38.0" prefWidth="50.0">
                     <graphic>
                         <ImageView fitHeight="34.0" fitWidth="37.0" pickOnBounds="true" preserveRatio="true">
-                            <Image url="@img/day-prev.png" />
+                            <Image url="@../img/day-prev.png" />
                         </ImageView>
                     </graphic>
                 </Button>
@@ -33,7 +33,7 @@
                 <Button fx:id="button_next" layoutX="10.0" layoutY="10.0" mnemonicParsing="false" onAction="#button_nextOnAction" prefHeight="38.0" prefWidth="50.0">
                     <graphic>
                         <ImageView fitHeight="34.0" fitWidth="37.0" pickOnBounds="true" preserveRatio="true">
-                            <Image url="@img/day-next.png" />
+                            <Image url="@../img/day-next.png" />
                         </ImageView>
                     </graphic>
                 </Button>
@@ -54,7 +54,7 @@
                 <Button fx:id="button_add" mnemonicParsing="false" onAction="#button_addOnAction" prefHeight="63.0" prefWidth="52.0">
                     <graphic>
                         <ImageView fitHeight="68.0" fitWidth="79.0" pickOnBounds="true" preserveRatio="true">
-                            <Image url="@img/add.png" />
+                            <Image url="@../img/add.png" />
                         </ImageView>
                     </graphic>
                 </Button>
@@ -86,7 +86,7 @@
                 <Button fx:id="button_delete" mnemonicParsing="false" onAction="#button_deleteOnAction" prefHeight="75.0" prefWidth="61.0">
                     <graphic>
                         <ImageView fitHeight="76.0" fitWidth="93.0" pickOnBounds="true" preserveRatio="true">
-                            <Image url="@img/delete.png" />
+                            <Image url="@../img/delete.png" />
                         </ImageView>
                     </graphic>
                 </Button>
