+++ src/main/resources/pl/gawor/taycknerdesktopclient/view/day_tracker/item_category.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<VBox fx:id="vbox_root" prefHeight="150.0" prefWidth="350.0" alignment="CENTER" onMouseClicked="#vbox_rootOnMouseClicked" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.ItemCategoryController">
+<VBox fx:id="vbox_root" prefHeight="150.0" prefWidth="350.0" alignment="CENTER" onMouseClicked="#vbox_rootOnMouseClicked" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.daytracker.ItemCategoryController">
     <Label fx:id="label_name" text="Work">
         <font>
           <Font size="24.0" />
