+++ src/main/resources/pl/gawor/taycknerdesktopclient/view/habit_tracker/item_habit.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<VBox fx:id="vbox_root" alignment="CENTER" onMouseClicked="#vbox_rootOnMouseClicked" prefHeight="150.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.ItemHabitController">
+<VBox fx:id="vbox_root" alignment="CENTER" onMouseClicked="#vbox_rootOnMouseClicked" prefHeight="150.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/11.0.2" xmlns:fx="http://javafx.com/fxml/1" fx:controller="pl.gawor.taycknerdesktopclient.controller.habittracker.ItemHabitController">
     <Label fx:id="label_name" text="Habit">
         <font>
           <Font size="24.0" />
